Modified ReadTSF and rfObject
ReadTSF modified in order to accept mixed reference impedances on the
format line.

ReadTSF now returns the S matrix on reference impedance asked in the
kwargs or 50 Ohm by default if not given, but as before on the reference
impedance found in the file.

ReadTSF now correctly reads 'variables' starting with a $-sign.

rfObject was modidied to also keep track of the 'variables' found in
touchstone files and returned via ReadTSF.